+++ src/msg.c
@@ -1991,7 +1991,7 @@ s32 get_msg_lines(s32 msgID) {
 }
 #endif
 
-void draw_msg(s32 msgID, s32 posX, s32 posY, s32 opacity, s32 palette, u8 style) {
+void draw_msg(s32 msgID, s32 posX, s32 posY, s32 opacity, s32 palette, u16 style) {
     MessagePrintState stackPrinter;
     MessagePrintState* printer;
     u16 bufferPos;
